Refactor visitors module into focused components
